+++ src/components/App.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react'
-import { Router, Route } from 'react-router-dom'
+import { ConnectedRouter } from 'connected-react-router';
+import { Route } from 'react-router-dom'
 import { Dispatch } from 'redux';
 import { connect } from 'react-redux';
 import { Action } from 'typescript-fsa';
@@ -10,21 +11,13 @@ import Eyecatch from './top/Eyecatch'
 import Top from './Top'
 import Detail from './Detail'
 import { authActions, workActions } from '../actions/actions';
+import history from '../history';
 import { AppState } from '../store';
 import { auth } from '../firebase'
 import Footer from './Footer';
 import About from './About';
 import SignIn from './SignIn';
-
 import ReactGA from 'react-ga';
-import { createBrowserHistory } from 'history'
-
-ReactGA.initialize('UA-107600889-2');
-const history = createBrowserHistory();
-history.listen(({ pathname }) => {
-    ReactGA.set({ page: pathname });
-    ReactGA.pageview(pathname);
-});
 
 interface AppActions {
     setUser: (user: firebase.User) => Action<firebase.User>;
@@ -56,7 +49,7 @@ const App: React.FC<AppProps> = (props: AppProps) => {
     return (
         <>
             <Helmet></Helmet>
-            <Router history={history}>
+            <ConnectedRouter history={history}>
                 <Eyecatch/>
                 <Header/>
                 <main>
@@ -78,7 +71,7 @@ const App: React.FC<AppProps> = (props: AppProps) => {
                     ))}
                 </main>
                 <Footer/>
-            </Router>
+            </ConnectedRouter>
         </>
     )
 }
